pref: profile image with cloudinary

+++ src/app/profile-image/profile-image.service.ts
@@ -1,6 +1,11 @@
 import { ERROR_CODE } from '../../interface'
 import { AppError } from '../../middleware'
-import { deleteFile, MESSAGE } from '../../utils'
+import {
+	MESSAGE,
+	uploadImageCloudinary,
+	updateImageCloudinary,
+	deleteImageCloudinary,
+} from '../../utils'
 import * as profileRepository from '../profile/profile.repository'
 
 import type {
@@ -13,28 +18,30 @@ import {
 } from './profile-image.mapper'
 import * as profileImageRepository from './profile-image.repository'
 
-const PORT = process.env.PORT ?? 3000
-const HOST = process.env.HOST ?? 'http://localhost'
-
 export const createProfileImage = async (
 	profileId: string,
 	file: Express.Multer.File,
-	fileUrl: string,
 ) => {
-	const { filename, size, path } = file
-	const profileImage: ICreateProfileImage = {
-		filename,
-		size,
-		path,
-		file_url: `${HOST}:${PORT}/profiles/files/${fileUrl}`,
-		profile_id: profileId,
-	}
 	const profile = await profileRepository.getProfile(profileId)
 	if (!profile) {
 		return new AppError(
 			ERROR_CODE.NOT_FOUND.code,
 			MESSAGE.ERROR.NOT_FOUND.PROFILE,
 		)
+	}
+	const upload = await uploadImageCloudinary(file.path, 'profile')
+	if (!upload) {
+		return new AppError(
+			ERROR_CODE.INTERNAL_SERVER_ERROR.code,
+			MESSAGE.ERROR.INTERNAL_SERVER_ERROR.CREATED.PROFILE_IMAGE,
+		)
+	}
+	const profileImage: ICreateProfileImage = {
+		public_id: upload.public_id,
+		filename: `${upload.public_id}.${upload.format}`,
+		size: upload.bytes,
+		file_url: upload.secure_url,
+		profile_id: profileId,
 	}
 	const createProfileImage =
 		await profileImageRepository.createProfileImage(profileImage)
@@ -50,15 +57,7 @@ export const createProfileImage = async (
 export const updateProfileImage = async (
 	profileImageId: string,
 	file: Express.Multer.File,
-	fileUrl: string,
 ) => {
-	const { filename, size, path } = file
-	const updateProfileImage: IUpdateProfileImage = {
-		filename,
-		size,
-		path,
-		file_url: `${HOST}:${PORT}/profiles/files/${fileUrl}`,
-	}
 	const profileImage =
 		await profileImageRepository.getProfileImage(profileImageId)
 	if (!profileImage) {
@@ -66,6 +65,18 @@ export const updateProfileImage = async (
 			ERROR_CODE.NOT_FOUND.code,
 			MESSAGE.ERROR.NOT_FOUND.PROFILE_IMAGE,
 		)
+	}
+	const upload = await updateImageCloudinary(file.path, profileImage.public_id)
+	if (!upload) {
+		return new AppError(
+			ERROR_CODE.INTERNAL_SERVER_ERROR.code,
+			MESSAGE.ERROR.INTERNAL_SERVER_ERROR.UPDATED.PROFILE_IMAGE,
+		)
+	}
+	const updateProfileImage: IUpdateProfileImage = {
+		filename: `${upload.public_id}.${upload.format}`,
+		size: upload.bytes,
+		file_url: upload.secure_url,
 	}
 	const profileImageUpdated = await profileImageRepository.updateProfileImage(
 		profileImageId,
@@ -77,7 +88,6 @@ export const updateProfileImage = async (
 			MESSAGE.ERROR.INTERNAL_SERVER_ERROR.UPDATED.PROFILE_IMAGE,
 		)
 	}
-	await deleteFile(profileImage.path)
 	return profileImageDTOMapper(profileImageUpdated)
 }
 
@@ -90,6 +100,13 @@ export const deleteProfileImage = async (profileImageId: string) => {
 			MESSAGE.ERROR.NOT_FOUND.PROFILE_IMAGE,
 		)
 	}
+	const deleteFile = await deleteImageCloudinary(profileImage.public_id)
+	if (!deleteFile) {
+		return new AppError(
+			ERROR_CODE.INTERNAL_SERVER_ERROR.code,
+			MESSAGE.ERROR.INTERNAL_SERVER_ERROR.DELETED.PROFILE_IMAGE,
+		)
+	}
 	const profileImageDeleted =
 		await profileImageRepository.deleteProfileImage(profileImageId)
 	if (!profileImageDeleted) {
@@ -98,6 +115,5 @@ export const deleteProfileImage = async (profileImageId: string) => {
 			MESSAGE.ERROR.INTERNAL_SERVER_ERROR.DELETED.PROFILE_IMAGE,
 		)
 	}
-	await deleteFile(profileImage.path)
 	return deleteProfileImageDTOMapper(profileImageDeleted)
 }
